Refactor ToDoListPage component to use custom hook for authentication and retrieve user's ToDo list
